+++ dashboard/utils.py
@@ -1,3 +1,4 @@
+from dashboard.forms import ExtractionScriptForm
 from dashboard.models import *
 
 def get_extracted_models(t):
@@ -13,14 +14,16 @@ def get_extracted_models(t):
         return (ExtractedCPCat, ExtractedListPresence)
 
     models = {
-        'Composition': one,
-        'Functional use': two,
-        'Chemical presence list': three
+        'CO': one,
+        'FU': two,
+        'CP': three
     }
     func = models.get(t, lambda: None)
     return func()
 
 def clean_dict(odict, model):
+    '''Cleans out key:value pairs that aren't in the model fields
+    '''
     return {f.name:odict[f.name]
             for f in model._meta.get_fields()
             if f.name in odict.keys()}
@@ -30,3 +33,15 @@ def update_fields(odict, model):
     for f in model._meta.get_fields():
         if f.name in odict.keys():
             setattr(model,f.name,odict[f.name])
+
+
+def include_extract_form(dg):
+    '''Returns the ExtractionScriptForm based on conditions of DataGroup
+    type as well as whether all records are matched, but not extracted
+    '''
+    if not dg.type in ['FU','CO','CP']:
+        return False
+    if dg.all_matched() and not dg.all_extracted():
+        return ExtractionScriptForm(dg_type=dg.type)
+    else:
+        return False
