+++ dashboard/views/data_document.py
@@ -1,4 +1,3 @@
-from django import forms
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect, get_object_or_404
 from dashboard.forms import (ExtractedTextForm, ExtractedCPCatForm,
