+++ dashboard/views/qa.py
@@ -1,7 +1,4 @@
-from django.forms import ModelForm, Form, BaseInlineFormSet, inlineformset_factory, TextInput, CharField
-
 from django.shortcuts import render, redirect, get_object_or_404
-
 from django.contrib.auth.decorators import login_required
 
 from dashboard.models import Script
